+++ lib/src/inputs/src/pickers/month/single.dart
@@ -307,7 +307,7 @@ class _ThemedMonthPickerState extends State<ThemedMonthPicker> {
                                           },
                                     child: Center(
                                       child: Text(
-                                        t('layrz.monthPicker.monthsShort.$index'),
+                                        t('layrz.monthPicker.months.$index'),
                                         style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                                               color: validateColor(color: decoration.color ?? Colors.white),
                                             ),